part of issue #373"

+++ libs/build_nceplibs.sh
@@ -94,6 +94,10 @@ if $MODULES; then
       module load sp
       module load ip2
       ;;
+    crtm)
+      module load hpc-$HPC_MPI
+      module load netcdf
+      ;;
     ip2)
       module load sp
       ;;
@@ -230,7 +234,7 @@ case $name in
     extraCMakeFlags="-DBUILD_POSTEXEC=OFF"
     ;;
   crtm)
-    URL="https://github.com/noaa-emc/emc_crtm"
+    URL="https://github.com/NOAA-EMC/crtm.git"
     ;;
   wgrib2)
     extraCMakeFlags="${STACK_wgrib2_cmake_opts:-}"
@@ -260,6 +264,9 @@ software=$name-$version
 if [[ ! -d $software ]]; then
   git clone $URL $software
   cd $software
+  if [[ "$name" == "crtm" ]]; then
+    version=release/REL-${install_as}_emc
+  fi
   git checkout $version
   git submodule update --init --recursive
 fi
@@ -269,19 +276,21 @@ cd ${HPC_STACK_ROOT}/${PKGDIR:-"pkg"}
 # Download CRTM fix files
 if [[ "$name" == "crtm" ]]; then
   if [[ ${STACK_crtm_install_fix:-} =~ [yYtT] ]]; then
-    if [[ ! -d crtm_fix-$version ]]; then
+    if [[ ! -d crtm_fix-${install_as} ]]; then
       crtm_tarball=fix_REL-${install_as}_emc.tgz
       rm -f $crtm_tarball
       $WGET ftp://ftp.ucar.edu/pub/cpaess/bjohns/$crtm_tarball
       tar xzf $crtm_tarball
-      mv fix crtm_fix-$version
+      mv fix crtm_fix-${install_as}
       rm -f $crtm_tarball
     fi
-    if [[ ! -f link_crtm_coeffs.sh ]]; then
-      $WGET https://raw.githubusercontent.com/NOAA-EMC/GSI/master/ush/link_crtm_coeffs.sh
-      sed -i'.backup' -e 's/LINK="ln -sf"/LINK="cp"/g' link_crtm_coeffs.sh
-      chmod +x link_crtm_coeffs.sh
-      rm -f link_crtm_coeffs.sh.backup
+    if [[ "${install_as}" == "2.3.0" ]]; then
+     if [[ ! -f link_crtm_coeffs.sh ]]; then
+       $WGET https://raw.githubusercontent.com/NOAA-EMC/GSI/master/ush/link_crtm_coeffs.sh
+       sed -i'.backup' -e 's/LINK="ln -sf"/LINK="cp"/g' link_crtm_coeffs.sh
+       chmod +x link_crtm_coeffs.sh
+       rm -f link_crtm_coeffs.sh.backup
+     fi
     fi
   fi
 fi
@@ -307,8 +316,22 @@ cd ${HPC_STACK_ROOT}/${PKGDIR:-"pkg"}
 # Install CRTM fix files
 if [[ "$name" == "crtm" ]]; then
   if [[ ${STACK_crtm_install_fix:-} =~ [yYtT] ]]; then
-    if [[ -d crtm_fix-$version ]]; then
-      ./link_crtm_coeffs.sh ./crtm_fix-$version $prefix/fix
+    if [[ -d crtm_fix-${install_as} ]]; then
+     if [[ "${install_as}" == "2.3.0" ]]; then
+       ./link_crtm_coeffs.sh ./crtm_fix-$version $prefix/fix
+     else
+       mkdir -p $prefix/fix
+       cp ./crtm_fix-${install_as}/ACCoeff/netcdf/* $prefix/fix
+       cp ./crtm_fix-${install_as}/AerosolCoeff/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/AerosolCoeff/netCDF/* $prefix/fix
+       cp ./crtm_fix-${install_as}/CloudCoeff/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/CloudCoeff/netCDF/* $prefix/fix
+       cp ./crtm_fix-${install_as}/EmisCoeff/*/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/EmisCoeff/*/*/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/SpcCoeff/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/SpcCoeff/netcdf/* $prefix/fix
+       cp ./crtm_fix-${install_as}/TauCoeff/ODPS/Big_Endian/* $prefix/fix
+     fi
     fi
   fi
 fi
